New SystemChannel Handling
Code Changes
- added SysChannel Command to manually define Channel for Systeminformations
  -> Introduced a new Handling and saving system for Systeminformationchannels
- Sorted some commands in the new packages k7bot.moderation.commands and k7bot.util.commands
- Optimized imports in Klassenserver7bbot.java

Bug Fixes
- Fixed some Help Command issues
- Fixed some JoinandLeaveListener Errors

+++ src/de/k7bot/moderation/commands/StopTimeoutCommand.java
@@ -1,7 +1,6 @@
-package de.k7bot.commands;
+package de.k7bot.moderation.commands;
 
 import de.k7bot.Klassenserver7bbot;
-
 import de.k7bot.commands.types.ServerCommand;
 import de.k7bot.util.PermissionError;
 import de.k7bot.util.SyntaxError;
@@ -9,6 +8,7 @@ import de.k7bot.util.SyntaxError;
 import java.time.OffsetDateTime;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
+
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Guild;
@@ -17,80 +17,75 @@ import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.TextChannel;
 import net.dv8tion.jda.api.exceptions.HierarchyException;
 
-public class kickCommand implements ServerCommand {
+public class StopTimeoutCommand implements ServerCommand {
 	public void performCommand(Member m, TextChannel channel, Message message) {
 		List<Member> ment = message.getMentionedMembers();
-
 		try {
-			String grund = message.getContentDisplay().substring(ment.get(0).getEffectiveName().length() + 8);
 
 			channel.sendTyping().queue();
 
-			if (m.hasPermission(Permission.KICK_MEMBERS)) {
+			if (m.hasPermission(Permission.MESSAGE_MANAGE)) {
 				if (ment.size() > 0) {
 					for (Member u : ment) {
-						onkick(m, u, channel, grund);
+						stopTimeout(m, u, channel);
 					}
 				}
 			} else {
 				PermissionError.onPermissionError(m, channel);
 			}
 		} catch (StringIndexOutOfBoundsException e) {
-			SyntaxError.oncmdSyntaxError(channel, "kick [@user] [reason]", m);
+			SyntaxError.oncmdSyntaxError(channel, "stoptimeout [@user]", m);
 		}
 	}
 
 	@Override
 	public String gethelp() {
-		String help = "Kickt den ausgewählten Nutzer vom Server und übermitelt den angegebenen Grund.\n - kann nur von Personen mit der Berechtigung 'Mitglieder kicken' ausgeführt werden!\n - z.B. kick @K7Bot [reason]";
-		return help;
+		return "Enttimeoutet den angegebenen Nutzer.\n - kann nur von Mitgliedern mit der Berechtigung 'Mitglieder kicken' ausgeführt werden!\n - z.B. [prefix]stoptimeout @member";
 	}
 
 	@Override
 	public String getcategory() {
-		String category = "Moderation";
-		return category;
+		return "Moderation";
 	}
 
-	public void onkick(Member requester, Member u, TextChannel channel, String grund) {
+	public void stopTimeout(Member requester, Member u, TextChannel channel) {
 		EmbedBuilder builder = new EmbedBuilder();
 		builder.setFooter("Requested by @" + requester.getEffectiveName());
 		builder.setTimestamp(OffsetDateTime.now());
 		builder.setThumbnail(u.getUser().getEffectiveAvatarUrl());
 		builder.setColor(16711680);
-		builder.setTitle("@" + u.getEffectiveName() + " was kicked");
+		builder.setTitle("@" + u.getEffectiveName() + " has been untimeouted");
 
 		StringBuilder strBuilder = new StringBuilder();
 		strBuilder.append("**User: **" + u.getAsMention() + "\n");
-		strBuilder.append("**Case: **" + grund + "\n");
 		strBuilder.append("**Requester: **" + requester.getAsMention() + "\n");
 
 		builder.setDescription(strBuilder);
 
 		Guild guild = channel.getGuild();
-		TextChannel system = guild.getSystemChannel();
+		TextChannel system = Klassenserver7bbot.INSTANCE.getsyschannell().getSysChannel(guild);
 
 		try {
-			u.kick().reason(grund).queue();
+			u.removeTimeout().queue();
 
-			if ((system = channel.getGuild().getSystemChannel()) != null) {
+			if (system != null) {
 
 				system.sendMessageEmbeds(builder.build()).queue();
 
 			}
 
-			if (system.getIdLong() != channel.getIdLong()) {
+			if (system != null && system.getIdLong() != channel.getIdLong()) {
 
 				channel.sendMessageEmbeds(builder.build()).complete().delete().queueAfter(20L, TimeUnit.SECONDS);
 
 			}
 
-			String action = "kick";
+			String action = "stoptimeout";
 			Klassenserver7bbot.INSTANCE.getDB().onUpdate(
 					"INSERT INTO modlogs(guildId, memberId, requesterId, memberName, requesterName, action, reason, date) VALUES("
 							+ channel.getGuild().getIdLong() + ", " + u.getIdLong() + ", " + requester.getIdLong()
 							+ ", '" + u.getEffectiveName() + "', '" + requester.getEffectiveName() + "', '" + action
-							+ "', '" + grund + "', '" + OffsetDateTime.now() + "')");
+							+ "', 'null', '" + OffsetDateTime.now() + "')");
 		} catch (HierarchyException e) {
 			PermissionError.onPermissionError(requester, channel);
 		}
